♻️ Give practice research one helper module
Every other theme keeps its home and show helpers in one module named
after the theme. Practice research was the last pair still split, so
the two merge with their constants and a single private section.
