+++ app/helpers/practice_research_helper.rb
@@ -1,9 +1,48 @@
 # frozen_string_literal: true
 
-module PracticeResearchShowHelper
+# Helpers for the practice research theme's home and show pages.
+module PracticeResearchHelper
+  NAMED_CONTRIBUTORS = 2
+  CARD_BLURB_LENGTH = 150
   VISIBLE_METADATA_ROWS = 8
   MAX_ROWS = 100
   DEFAULT_ROWS = 20
+
+  def pr_home
+    theme_home(PracticeResearch::HomepagePresenter)
+  end
+
+  def pr_featured_researcher?
+    @featured_researcher&.value.present?
+  end
+
+  def pr_card_blurb(presenter, length: CARD_BLURB_LENGTH)
+    text = Array(presenter.try(:description)).first
+    return if text.blank?
+
+    truncate(pr_plain_text(text), length:, separator: ' ')
+  end
+
+  def pr_plain_text(html)
+    Nokogiri::HTML.fragment(html.to_s.gsub(%r{</?[a-zA-Z][^>]*>}, ' ')).text.squish
+  end
+
+  def pr_thumbnail?(presenter)
+    path = presenter.thumbnail_path
+
+    path.is_a?(String) && path.present? && path != Hyrax::ThumbnailPathService.default_image
+  end
+
+  def pr_contributor_summary(document)
+    names = pr_contributor_names(document)
+    return if names.empty?
+
+    rest = names.size - NAMED_CONTRIBUTORS
+    return names.to_sentence if rest < 1
+
+    t('practice_research.homepage.featured.contributors_and_others',
+      names: names.first(NAMED_CONTRIBUTORS).join('; '), count: rest)
+  end
 
   def pr_show_panes(presenter)
     @pr_show_panes ||= [].tap do |panes|
@@ -75,6 +114,17 @@ module PracticeResearchShowHelper
   end
 
   private
+
+  def pr_contributor_names(document)
+    rows = begin
+             JSON.parse(Array(document['participants_json_ss']).first.to_s)
+           rescue JSON::ParserError
+             []
+           end
+    names = Array(rows).filter_map { |row| row['name'] if row.is_a?(Hash) }.compact_blank
+
+    names.presence || Array(document['creator_tesim']).compact_blank
+  end
 
   def pr_rows
     theme_positive_param(:rows, DEFAULT_ROWS, MAX_ROWS)
